feat: Add new API endpoint for agent flags
Adds a new API endpoint at `/v1/flags` to retrieve the agent flags. This
is a temporary endpoint while the APIs are being updated, and will be
used by the agent to fetch the flags. The existing `/flags` endpoint
will continue to be used by the library, and the
`/environment/{environmentId}/flags` endpoint will be used by the
frontend.